move code component EmptyState reliable news
