+++ src/pages/Home/components/News/index.view.js
@@ -1,6 +1,4 @@
-import React from 'react';
 import styled from 'styled-components';
-import { ReactComponent as EmptyIcon } from './assets/subtract.svg';
 
 const Wrapper = styled.div`
   padding: 16px;
@@ -20,39 +18,4 @@ const WrapperCards = styled.div`
   margin-top: 15px;
 `;
 
-const EmptyWrapper = styled.div`
-  height: 209px;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  ${({ theme }) => theme.typography.body()}
-`;
-
-const EmptyTitle = styled.div`
-  margin: 15px auto 4px;
-`;
-
-const EmptyDescription = styled.div`
-  max-width: 307px;
-  ${({ theme }) => theme.typography.bodySmallHighlight()};
-  text-align: center;
-`;
-
-const EmptyState = () => {
-  return (
-    <EmptyWrapper>
-      <EmptyIcon />
-      <EmptyTitle>ยังไม่มีข้อมูล</EmptyTitle>
-      <EmptyDescription>
-        กำลังประมวลผล และคัดกรองข่าวจากชุมชน <br />
-        คุณสามารถเป็นส่วนนึงของการคัดกรองข่าว
-        <br />
-        โดยการโหวตความน่าเชื่อถือของข่าวได้ ในส่วนของ “ข่าวทั้งหมด”
-      </EmptyDescription>
-    </EmptyWrapper>
-  );
-};
-
-export { Wrapper, ButtonsWrapper, WrapperCards, EmptyState };
+export { Wrapper, ButtonsWrapper, WrapperCards };
